+++ pytorch_lightning/trainer/evaluation_loop.py
@@ -11,31 +11,31 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from typing import Dict, List, Union
+from typing import Any, Dict, List, Optional, Tuple, Union
 
-import torch
+from torch.utils.data import DataLoader
 
+import pytorch_lightning as pl
 from pytorch_lightning.core.step_result import Result
 from pytorch_lightning.trainer.states import TrainerState
 from pytorch_lightning.trainer.supporters import PredictionCollection
-from pytorch_lightning.utilities.apply_func import apply_to_collection
 from pytorch_lightning.utilities.model_helpers import is_overridden
 from pytorch_lightning.utilities.signature_utils import is_param_in_hook_signature
+from pytorch_lightning.utilities.types import EPOCH_OUTPUT, STEP_OUTPUT
 from pytorch_lightning.utilities.warnings import WarningCache
 
 
 class EvaluationLoop(object):
 
-    def __init__(self, trainer):
-        self.trainer = trainer
-        self.outputs = []
-        self.step_metrics = []
-        self.predictions = None
-        self.max_batches = None
+    def __init__(self, trainer: 'pl.Trainer'):
+        self.trainer: 'pl.Trainer' = trainer
+        self.outputs: EPOCH_OUTPUT = []
+        self.predictions: Optional[PredictionCollection] = None
+        self.max_batches: Optional[List[Union[int, float]]] = None
         self.warning_cache = WarningCache()
-        self.num_dataloaders = None
+        self.num_dataloaders: Optional[int] = None
 
-    def on_trainer_init(self):
+    def on_trainer_init(self) -> None:
         self.trainer.num_sanity_val_batches = []
         self.trainer.num_test_batches = []
         self.trainer.num_val_batches = []
@@ -49,7 +49,7 @@ class EvaluationLoop(object):
         # when true, print evaluation results in .validate() and .test()
         self.trainer.verbose_evaluate = True
 
-    def get_evaluation_dataloaders(self):
+    def get_evaluation_dataloaders(self) -> Tuple[Optional[List[DataLoader]], List[Union[int, float]]]:
         model = self.trainer.lightning_module
 
         # select dataloaders
@@ -72,30 +72,30 @@ class EvaluationLoop(object):
             dataloaders = self.trainer.val_dataloaders
         return dataloaders, max_batches
 
-    def should_skip_evaluation(self, max_batches):
+    def should_skip_evaluation(self, max_batches: List[Union[int, float]]) -> bool:
         return sum(max_batches) == 0
 
-    def on_evaluation_start(self, *args, **kwargs):
+    def on_evaluation_start(self, *args: Any, **kwargs: Any) -> None:
         if self.trainer.testing:
             self.trainer.call_hook('on_test_start', *args, **kwargs)
         else:
             self.trainer.call_hook('on_validation_start', *args, **kwargs)
 
-    def on_evaluation_model_eval(self, *_, **__):
+    def on_evaluation_model_eval(self) -> None:
         model_ref = self.trainer.lightning_module
         if self.trainer.testing:
             model_ref.on_test_model_eval()
         else:
             model_ref.on_validation_model_eval()
 
-    def on_evaluation_model_train(self, *_, **__):
+    def on_evaluation_model_train(self) -> None:
         model_ref = self.trainer.lightning_module
         if self.trainer.testing:
             model_ref.on_test_model_train()
         else:
             model_ref.on_validation_model_train()
 
-    def on_evaluation_end(self, *args, **kwargs):
+    def on_evaluation_end(self, *args: Any, **kwargs: Any) -> None:
         if self.trainer.testing:
             self.trainer.call_hook('on_test_end', *args, **kwargs)
         else:
@@ -105,14 +105,14 @@ class EvaluationLoop(object):
             # summarize profile results
             self.trainer.profiler.describe()
 
-    def reload_evaluation_dataloaders(self):
+    def reload_evaluation_dataloaders(self) -> None:
         model = self.trainer.lightning_module
         if self.trainer.testing:
             self.trainer.reset_test_dataloader(model)
         else:
             self.trainer.reset_val_dataloader(model)
 
-    def setup(self, model, max_batches, dataloaders):
+    def setup(self, max_batches: List[Union[int, float]], dataloaders: List[DataLoader]) -> None:
         # bookkeeping
         self.outputs = []
         self.predictions = PredictionCollection(self.trainer.global_rank, self.trainer.world_size)
@@ -123,9 +123,8 @@ class EvaluationLoop(object):
 
         self.max_batches = max_batches
         self.num_dataloaders = self._get_num_dataloaders(dataloaders)
-        self._predictions = [[] for _ in range(self.num_dataloaders)]
 
-    def on_evaluation_epoch_start(self, *args, **kwargs):
+    def on_evaluation_epoch_start(self, *args: Any, **kwargs: Any) -> None:
         self.trainer.call_hook('on_epoch_start', *args, **kwargs)
 
         if self.trainer.testing:
@@ -133,7 +132,7 @@ class EvaluationLoop(object):
         else:
             self.trainer.call_hook('on_validation_epoch_start', *args, **kwargs)
 
-    def _build_args(self, batch, batch_idx, dataloader_idx):
+    def _build_args(self, batch: Any, batch_idx: int, dataloader_idx: int) -> List[Union[Any, int]]:
         # make dataloader_idx arg in validation_step optional
         args = [batch, batch_idx]
 
@@ -147,15 +146,18 @@ class EvaluationLoop(object):
 
         return args
 
-    def _get_num_dataloaders(self, dataloaders):
+    def _get_num_dataloaders(self, dataloaders: Optional[List[DataLoader]]) -> int:
         # case where user does:
         # return dl1, dl2
-        length = len(dataloaders)
-        if len(dataloaders) > 0 and isinstance(dataloaders[0], (list, tuple)):
-            length = len(dataloaders[0])
-        return length
+        if dataloaders is not None:
+            length = len(dataloaders)
+            if len(dataloaders) > 0 and isinstance(dataloaders[0], (list, tuple)):
+                length = len(dataloaders[0])
+            return length
+        else:
+            return 0
 
-    def evaluation_step(self, batch, batch_idx, dataloader_idx):
+    def evaluation_step(self, batch: Any, batch_idx: int, dataloader_idx: int) -> Optional[STEP_OUTPUT]:
         # configure args
         args = self._build_args(batch, batch_idx, dataloader_idx)
 
@@ -174,20 +176,19 @@ class EvaluationLoop(object):
         # capture any logged information
         self.trainer.logger_connector.cache_logged_metrics()
         # track batch size for weighted average
-        is_result_obj = isinstance(output, Result)
-        if is_result_obj:
+        if isinstance(output, Result):
             output.track_batch_size(batch)
 
         return output
 
-    def evaluation_step_end(self, *args, **kwargs):
+    def evaluation_step_end(self, *args: Any, **kwargs: Any) -> Optional[STEP_OUTPUT]:
         if self.trainer.testing:
             output = self.trainer.call_hook('test_step_end', *args, **kwargs)
         else:
             output = self.trainer.call_hook('validation_step_end', *args, **kwargs)
         return output
 
-    def evaluation_epoch_end(self, outputs):
+    def evaluation_epoch_end(self, outputs: EPOCH_OUTPUT) -> None:
         # unset dataloder_idx in model
         self.trainer.logger_connector.evaluation_epoch_end()
 
@@ -207,41 +208,7 @@ class EvaluationLoop(object):
         # capture logging
         self.trainer.logger_connector.cache_logged_metrics()
 
-    def __gather_epoch_end_eval_results(self, outputs):
-        eval_results = []
-        for epoch_output in outputs:
-            result = epoch_output[0].__class__.gather(epoch_output)
-            eval_results.append(result)
-
-        # with 1 dataloader don't pass in a list
-        if len(eval_results) == 1:
-            eval_results = eval_results[0]
-        return eval_results
-
-    def __auto_reduce_result_objs(self, outputs):
-        # outputs has a list of results per dataloader
-        eval_results = []
-        for dl_output in outputs:
-            result = dl_output[0]
-            result = result.__class__.reduce_on_epoch_end(dl_output)
-            eval_results.append(result)
-
-        return eval_results
-
-    def on_predict_epoch_end(self):
-        if self.trainer._progress_bar_callback is not None:
-            self.trainer._progress_bar_callback.on_test_end(self.trainer, self.trainer.lightning_module)
-
-        results = self._predictions
-
-        def _convert_to_numpy(v):
-            return v.cpu().numpy()
-
-        results = apply_to_collection(results, torch.Tensor, _convert_to_numpy)
-
-        return results, None
-
-    def on_evaluation_batch_start(self, batch, batch_idx, dataloader_idx):
+    def on_evaluation_batch_start(self, batch: Any, batch_idx: int, dataloader_idx: int) -> None:
         # set dataloader_idx to model and track batch_size
         self.trainer.logger_connector.on_evaluation_batch_start(batch, dataloader_idx, self.num_dataloaders)
 
@@ -250,7 +217,13 @@ class EvaluationLoop(object):
         else:
             self.trainer.call_hook('on_validation_batch_start', batch, batch_idx, dataloader_idx)
 
-    def on_evaluation_batch_end(self, output, batch, batch_idx, dataloader_idx):
+    def on_evaluation_batch_end(
+        self,
+        output: Optional[STEP_OUTPUT],
+        batch: Any,
+        batch_idx: int,
+        dataloader_idx: int,
+    ) -> None:
         if self.trainer.testing:
             self.trainer.call_hook('on_test_batch_end', output, batch, batch_idx, dataloader_idx)
         else:
@@ -259,11 +232,10 @@ class EvaluationLoop(object):
         # store predicitons if do_write_predictions and track eval loss history
         self.store_predictions(output, batch_idx, dataloader_idx)
 
-    def store_predictions(self, output, batch_idx, dataloader_idx):
+    def store_predictions(self, output: Optional[STEP_OUTPUT], batch_idx: int, dataloader_idx: int) -> None:
         # Add step predictions to prediction collection to write later
-        if output is not None:
-            do_write_predictions = isinstance(output, Result) and self.trainer.testing
-            if do_write_predictions:
+        if output is not None and self.predictions is not None:
+            if isinstance(output, Result) and self.trainer.testing:
                 self.predictions.add(output.pop('predictions', None))
 
         # track debug metrics
@@ -296,29 +268,21 @@ class EvaluationLoop(object):
 
         self.trainer.call_hook('on_epoch_end')
 
-    def log_evaluation_step_metrics(self, output, batch_idx):
+    def log_evaluation_step_metrics(self, batch_idx: int) -> None:
         if self.trainer.sanity_checking:
             return
 
-        step_log_metrics = {}
-        step_pbar_metrics = {}
-
-        self.__log_result_step_metrics(step_log_metrics, step_pbar_metrics, batch_idx)
-
-    def __log_result_step_metrics(self, step_log_metrics, step_pbar_metrics, batch_idx):
         cached_results = self.trainer.logger_connector.cached_results
-        cached_batch_pbar_metrics, cached_batch_log_metrics = cached_results.update_logger_connector()
+        if cached_results is not None:
+            cached_batch_pbar_metrics, cached_batch_log_metrics = cached_results.update_logger_connector()
 
-        step_log_metrics.update(cached_batch_log_metrics)
-        step_pbar_metrics.update(cached_batch_pbar_metrics)
+            if len(cached_batch_log_metrics) > 0:
+                # make the metrics appear as a different line in the same graph
+                metrics_by_epoch = {}
+                for k, v in cached_batch_log_metrics.items():
+                    metrics_by_epoch[f'{k}/epoch_{self.trainer.current_epoch}'] = v
 
-        if len(step_log_metrics) > 0:
-            # make the metrics appear as a different line in the same graph
-            metrics_by_epoch = {}
-            for k, v in step_log_metrics.items():
-                metrics_by_epoch[f'{k}/epoch_{self.trainer.current_epoch}'] = v
+                self.trainer.logger_connector.log_metrics(metrics_by_epoch, {}, step=batch_idx)
 
-            self.trainer.logger_connector.log_metrics(metrics_by_epoch, {}, step=batch_idx)
-
-        if len(step_pbar_metrics) > 0:
-            self.trainer.logger_connector.add_progress_bar_metrics(step_pbar_metrics)
+            if len(cached_batch_pbar_metrics) > 0:
+                self.trainer.logger_connector.add_progress_bar_metrics(cached_batch_pbar_metrics)